moved md related data handlers to data.md sub-package
