+++ chainsaw/data/md/featurization/angles.py
@@ -28,9 +28,9 @@ from mdtraj.geometry.dihedral import (indices_phi,
                                       )
 import mdtraj
 
-from chainsaw.data.featurization._base import Feature
-from chainsaw.data.featurization.util import (_hash_numpy_array,
-                                                        hash_top, _describe_atom)
+from chainsaw.data.md.featurization._base import Feature
+from chainsaw.data.md.featurization.util import (_hash_numpy_array,
+                                                 hash_top, _describe_atom)
 import numpy as np
