+++ chainsaw/tests/test_api_source.py
@@ -16,14 +16,13 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-
-
 from __future__ import absolute_import
+
 import unittest
 import os
 import tempfile
 
-from chainsaw.data import MDFeaturizer
+from chainsaw.data.md import MDFeaturizer
 from logging import getLogger
 import chainsaw.api as api
 import numpy as np
