+++ chainsaw/tests/test_featurereader_and_tica.py
@@ -22,17 +22,18 @@ Test feature reader and Tica with a set of cosine time series.
 @author: Fabian Paul
 '''
 
+from __future__ import absolute_import
 from __future__ import print_function
 
-from __future__ import absolute_import
-import unittest
 import os
 import tempfile
-import numpy as np
-import mdtraj
-from chainsaw import api
-from chainsaw.data.feature_reader import FeatureReader
+import unittest
 from logging import getLogger
+
+import mdtraj
+import numpy as np
+from chainsaw import api
+from chainsaw.data.md.feature_reader import FeatureReader
 from six.moves import range
 
 log = getLogger('chainsaw.'+'TestFeatureReaderAndTICA')
